refactor: consolidate test scaffolding, shrink stale ratchet baselines, merge misc settings slices
Three structural-debt cleanups, each verified byte/live-source accurate
before applying:

- Extract the byte-identical stripComments()/sourceFilesUnder() helper
  bodies duplicated across 4 architecture-ratchet test files into
  test-kernel/support/repoScan.ts as stripComments() and
  productionFilesUnder(); dependencyDirection.vitest.ts keeps its own
  sourceFilesUnder() (genuinely different 2-arg signature).
- Shrink host-agent-import-baseline.json and
  shared-schemas-deep-import-baseline.json to drop 12 tracked
  specifiers/entries with zero remaining occurrences in source, and add
  2 legitimately new @shared/schemas/agent gratuitous imports the
  baseline had drifted behind (src/model/ToolDefinition.ts,
  src/tools/executionFormatters.ts) so the ratchet reflects live state
  instead of stale slack. Both ratchets are upper-bound/count checks
  only, so this cannot break CI.
- Merge six 10-18-line single-command settingsView slice files
  (agentSkillsSettingsSlice, telemetrySettingsSlice, goalSlice,
  agentTeamsSlice, toolDashboardSlice, multiAgentSlice) into one
  miscSettingsSlice.ts; none had cross-file coupling beyond
  ../settingsState. modelSelectionSlice.ts and approvalSettingsSlice.ts
  are also single-command but stay standalone since each owns a named
  state section rather than a one-off toggle — documented in
  messageDispatcher.ts's header comment.

shortlist, adversarial net-gain verification) scoped to consolidation
and native-method opportunities; three other candidates it surfaced
(vi.mock('vscode') convergence, getExecutionStore test-mock
convergence, an agentConfig/OAuth-timeout/createRunScope micro-batch)
came back WASH on adversarial verification and were dropped.
