+++ config/ratchets/host-agent-import-baseline.json
@@ -72,12 +72,10 @@
       "@agent/index",
       "@agent/index/AgentDirectorySync",
       "@agent/index/platformAgentDirectories",
-      "@agent/modelHandlers/openai/modelHandlerOpenAI",
       "@agent/remote/remoteAgentConfigClient",
       "@agent/review/reviewDiff",
       "@agent/review/reviewIssues",
       "@agent/runtime/AgentFlowResult",
-      "@agent/runtime/agentLoad",
       "@agent/runtime/agentShutdown",
       "@agent/runtime/detachSubagentsOnStop",
       "@agent/runtime/ExecutionHandle",
@@ -94,13 +92,10 @@
       "@agent/runtime/SessionHandle",
       "@agent/runtime/SessionResumeRetrieval",
       "@agent/runtime/terminalResultToast",
-      "@agent/runtime/textConnection",
       "@agent/runtime/textEnhancement",
       "@agent/storage",
-      "@agent/storage/detectWaitingStreams",
       "@agent/templates/agentTemplateRenderer",
-      "@agent/trace",
-      "@agent/utils/agentConfigToTaskState"
+      "@agent/trace"
     ]
   }
 }
