+++ src/test-kernel/architecture/sessionFactAmbientHelperRetirement.vitest.ts
@@ -6,8 +6,9 @@ import { resolve } from 'node:path';
 import { describe, expect, it } from 'vitest';
 
 import {
+  productionFilesUnder,
   REPO_ROOT,
-  sourceFilesUnder as sharedSourceFilesUnder,
+  stripComments,
 } from '../support/repoScan';
 
 const ALLOWED_PRODUCTION_REFERENCES = [] as const;
@@ -21,25 +22,6 @@ const SCAN_ROOTS = [
 
 const RETIRED_AMBIENT_HELPER_SYMBOL = /\bemitRuntimeEvent\b/;
 
-function sourceFilesUnder(root: string): string[] {
-  return sharedSourceFilesUnder(resolve(REPO_ROOT, root), {
-    missingDirReturnsEmpty: true,
-    repoRelative: true,
-    excludeTestKernel: true,
-  });
-}
-
-function stripComments(source: string): string {
-  return source
-    .replaceAll(/\/\*[\s\S]*?\*\//g, '')
-    .split('\n')
-    .map((line) => {
-      const commentStart = line.indexOf('//');
-      return commentStart === -1 ? line : line.slice(0, commentStart);
-    })
-    .join('\n');
-}
-
 function referencesEmitRuntimeEvent(file: string): boolean {
   const source = stripComments(readFileSync(resolve(REPO_ROOT, file), 'utf8'));
   return RETIRED_AMBIENT_HELPER_SYMBOL.test(source);
@@ -47,7 +29,7 @@ function referencesEmitRuntimeEvent(file: string): boolean {
 
 describe('ambient session-fact helper retirement', () => {
   it('keeps the retired ambient session-fact helper out of production code', () => {
-    const references = SCAN_ROOTS.flatMap(sourceFilesUnder)
+    const references = SCAN_ROOTS.flatMap(productionFilesUnder)
       .filter(referencesEmitRuntimeEvent)
       .toSorted();
 
@@ -56,7 +38,7 @@ describe('ambient session-fact helper retirement', () => {
 
   it('actually scans the production source roots', () => {
     const scanned = SCAN_ROOTS.reduce(
-      (total, root) => total + sourceFilesUnder(root).length,
+      (total, root) => total + productionFilesUnder(root).length,
       0,
     );
     expect(scanned).toBeGreaterThan(100);
